Added more things
Fixed the power ups system, changing the performance, with the switch statement. Also added a new Power-Up: GOD MODE: Makes your player has the height of the screen for a 3 seconds.
Also, i replaced the raylib .dll and .lib, to a x32 bit version and created a Release with the game at this time of the development.